requirements.txt
scipy==1.4.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
easydict==1.9
numpy==1.18.1
Pillow==7.1.2


